refactor chat page with hooks and layout
